fix client render
